pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["renato <renato@teste.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
django = "^5.1.7"
pytest = "^8.3.4"
factory-boy = "^3.3.1"
djangorestframework = "^3.15.2"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--reuse-db"
DJANGO_SETTINGS_MODULE = "bookstore.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]